Moving validation strings and constants to core instead of modeler
