+++ AutoRest/AutoRest.Core/AutoRest.Core.csproj
@@ -86,6 +86,7 @@
     <Compile Include="Utilities\MemoryFileSystem.cs" />
     <Compile Include="Validation\IValidator.cs" />
     <Compile Include="Validation\ValidationException.cs" />
+    <Compile Include="Validation\ValidationExceptionConstants.cs" />
     <Compile Include="Validation\ValidationMessage.cs" />
   </ItemGroup>
   <ItemGroup>
