+++ AutoRest/Modelers/Swagger/AutoRest.Modeler.Swagger.csproj
@@ -73,7 +73,6 @@
     <Compile Include="Model\Info.cs" />
     <Compile Include="Model\Tag.cs" />
     <Compile Include="Model\TransferProtocolScheme.cs" />
-    <Compile Include="ValidationExceptionConstants.cs" />
     <Compile Include="Validators\Attributes\AnonymousTypes.cs" />
     <Compile Include="Validators\Attributes\OperationIdSingleUnderscore.cs" />
     <Compile Include="Validators\Attributes\Rule.cs" />
